floating blocks don't attach physics after operator interactions if mounted in a container
fixes #3

note: this can lead to confusing mount behaviour if one object is mounted and another is dragged onto it, meaning the second (if operator popup is cancelled) can be left "hanging"  on the container (it's actually mounted). Likely best resolved by implementing container auto-layout or avoiding through problem design.